Add support for qualifyingmiles.com calculation.
Per a request on FlyerTalk, this commit adds support for calculating earnings
on Delta PNRs with qualifyingmiles.com. I kept the link to
skymilescalculator.com as well in case folks want to use that.

I don't have an Aeromexico or Virgin Atlantic PNR to test with, so I did not
update those airlines' earnings calculations.